+++ src/CommonMarkExtrasExtension.php
@@ -15,14 +15,16 @@ use League\CommonMark\ConfigurableEnvironmentInterface;
 use League\CommonMark\Ext\Autolink\AutolinkExtension;
 use League\CommonMark\Ext\SmartPunct\SmartPunctExtension;
 use League\CommonMark\Ext\Strikethrough\StrikethroughExtension;
+use League\CommonMark\Ext\TaskList\TaskListExtension;
 use League\CommonMark\Extension\ExtensionInterface;
 
 final class CommonMarkExtrasExtension implements ExtensionInterface
 {
     public function register(ConfigurableEnvironmentInterface $environment)
     {
+        $environment->addExtension(new AutolinkExtension());
         $environment->addExtension(new SmartPunctExtension());
         $environment->addExtension(new StrikethroughExtension());
-        $environment->addExtension(new AutolinkExtension());
+        $environment->addExtension(new TaskListExtension());
     }
 }